refactor(frontend): move shop appearance type and parser out of admin pages
- ShopAppearanceFormState -> src/types/shopAppearance.ts
- parseAppearanceFormDefaults -> src/utils/shopAppearance.ts
- shopSettingsForm.ts only exports buildAppearancePatch
- useAdminShopBranding imports shared utils (no page-layer coupling)
