+++ frontend/src/types/shopAppearance.ts
